ref/quarcc-trading-engine/python_client/requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0
